Consider only one review event per user (#35)
Previously, if a user left multiple approval comments or left a comment
and a review, their approval would count twice. Now, review candidates
are deduplicated so that only the most recent for each user is
considered.

As a result, add the LastModified field to test data so that generated
strings have a consistent order.