Correctly handle dismissed GitHub reviews
When a review is dismissed, remove it from the review list as though it
never happened. This should be equivalent to a user deleting an approval
comment, although unlike in that case, we don't audit for this event.
Because GitHub provides the ability to restrict review dismissal, this
seems safe.