Fix invalidate_on_push, use GraphQL for timeline (#37)
Previously, you could bypass the invalidate_on_push option by faking
commit times. The v4 API exposes push times as a property, so switch to
using that for timeline requests. This also simplifies loading commits
and removes the need for the order field.

For testing, make ResponsePlayer more flexible by introducing a matcher
interface.  Add an exact path matcher for the old behavior and a GraphQL
matcher that parses queries and responds if the query contains a given
node path prefix.