Fix login strings for bot accounts (#44)
In the V3 API, the login strings for bot accounts had a "[bot]" suffix,
which may be used in policy definitions. In the V4 API, bot accounts
have their own type, so the login string does not include this suffix.
Add the suffix in our code to maintain compatibility with existing
policy definitions.