Remove usage of pull request timeline API (#54)
We originally used this API to have consistent ordering between
comments, commits, and reviews. However, this didn't respect commit push
times, so we switched to using timestamps for ordering with the switch
to GraphQL. Later, we discovered that the timeline API has a random
delay before commits, particularly from forks, appear in the result.
This causes the bot to evaluate the status for the wrong commit,
breaking the invalidate_on_push option.

Instead, get commits, comments, and reviews from the pull request API,
which should be up-to-date when we respond to a pull_request hook.
Because all three of these sources are used in most evaluations,
continue to load them with a single query instead of making separate
queries for each type.