Verify that the pull request head commit exists
Some GitHub APIs have a delay propagating information about new commits,
particularly when the commits come from forks. While I believe the APIs
we currently use are safe, if they are not, the policy should fail
closed with an error instead of potentially failing open, as it does
currently.

Also reorganize some test data so that it is more consistent, making it
easier for tests to pass this new check.