Fix invalidate_on_push for fork pull requests
GitHub does not propagate all commit information from forks to the
upstream repository when creating a pull request. In particular,
pushedDate is always null when listing commits in a fork via the pull
request APIs. To fix this, always list commits directly from the head
repository.

As a consequence, this also removes the 250 commit limit, although any
PR with more than 250 commits is likely to also hit the 300 file limit.

Finally, reorganize the test data and construct the default test PR in
code instead of optionally loading it from a mock request.