Load full PR if it is missing required fields
The PR objects included in webhook payloads are oddly inconsistent: the
pull_request_review payload does not include the "commits" field, which
is require for policy-bot to function. Instead of loading PRs in each
handler, modify the context to load the PR if the one it was given is
missing any required fields. To make this harder to mess up, use a new
pull request object internally based on the GraphQL query.