Load pull request at context creation
Introduce a Locator type to hold the identifying information for a PR
and an optional PR object. If the PR is missing or is not complete, make
an API request at creation time for the missing data. Now that this data
is guaranteed, remove the error return values from methods that only use
PR data.

Also refactor how the base handler evaluates PRs to reduce the number of
arguments and to pass pull.Context values instead of PR objects.