Load full PR if it is missing required fields (#68)
The PR objects included in webhook payloads are oddly inconsistent: the
pull_request_review payload does not include the "commits" field, which
is required for policy-bot to function. Modify the context to load
additional pull request details on demand if it is given an object
missing required fields.

Now that these values are guaranteeded to exist, remove the error return
values from context methods that only use these values. This also
refactors how the base handler evaluates PRs to reduce the number of
arguments and to standardize on passing pull.Context values instead of
PR objects.