Return to loading commits from the pull request
This is the only way to get only the commits included in the PR without
performing our own comparisons against the commits on the target branch.
This property is important to correctly handle merge commits.

In order to load pushedDate values from fork commits, conditionally make
a second query listing commit history and match these commits with the
ones in the pull request.

To avoid additional hazards of commit loading, simplify the detection of
update merges by relying on the principle that GitHub will automatically
exclude any commits already on the target branch from the PR listing.
This removes the need to load target branch commits.

Finally, change the pushed date backfill logic to walk the parent commit
chain instead of using a specific order that differs between APIs.