Support policy files up to 100 MB
Previous usage of Git Repos API only supports policy files with max
size of 1 MB. This adds a fall-back to use the Git Data API if the
initial call fails as the file is too large.

The DownloadContents method could be used in place of GetContents,
but since this code is called very often and DownloadContents makes
an additional API call to list the tree to acquire the file SHA,
and this is an edge case, I have implemented it as a fallback instead.