Increase the number of retries for missing dates
Internally, the existing timeouts are not always sufficient. Adjust the
retry loop to use exponential backoff with more attempts, resulting in a
total of 15s between the first and last attempts instead of the 3s in
the old implementation. Note that GitHub hooks have a 30s timeout, so we
can't go too much higher on the retry counts.