Avoid retries for pushedDate loading (#154)
Previously, this fallback API was only used for pull requests from forks
and we would retry the primary API call for other pull requests. But in
practice, we had to retry for up to 15 seconds before success and even
then saw occasional failures. Using the fallback method (the commit
history API) should be faster and more correct in all cases (the ideal
case still being that we get accurate information on the first call.)

This also adds a new exit condition to the fallback loop to avoid
loading all commits in history under certain error cases.