Unmarshal regexps directly from configuration (#157)
Introduce the common.Regexp type as a simple wrapper for regexp.Regexp
that adds unmarshalling methods for JSON and YAML. This simplifies code
using these expressions and ensures that invalid regular expressions are
caught early and are detectable with the validation API.

The downside is that invalid expressions now break the loading
processes, so we can't report multiple errors at once. We don't do this
now, but if we want to in the future, another refactoring will be
required.