Add ignore_commits_by option (#156)
This option ignores all commits that have both an author and
contributor matching the conditions in the block. These commits
effectively do not exist to policy-bot, meaning they don't influence
invalidation on push or contributor detection.

This can be useful to ignore certain automated commits or commits from
trusted user groups. Note that commit authorship is easy to forge, so
this is not a secure option unless paired with a pre-receive hook that
enforces accurate commit attribution.