Seed RNG using pull request creation time
This means that reviewer selection for any given PR will be consistent
over the life of the PR. I think creation time will be granular enough
that even bulk-opened PRs should get different reviewers, but we could
mix in the PR number or ID if we need more variance between PRs.

This only works if the non-random aspects of reviewer selection are
deterministic (e.g. no map iteration), which I'll have to check next.