Fix post-login redirect without a saved path
Normally, the session contains a redirect target containing the path you
visited before being redirected to authenticate. In some cases, it's
possible to end up at the login page without a target in the session,
which should redirect you to the base path. Unfortunately, we used the
empty string, which redirected to an unmapped route.