Revert back to go-github v32
In v33, the library removed some preview media type headers that are
still required in GitHub Enterprise, although they are no longer
required on GitHub.com. Until there's a GHE release that drops the
requirement for the preview media types, we'll need to stick with the
older version.

Since we otherwise weren't taking advantage of new features or changes,
this should be fine.