Fix conflict between direct and inherited perms
Previously, if the aggregate permission for a collaborator was higher
than a permission granted directly, the direct permission was discarded
and the user would not be selected for reviews. This was particularly
bad for org admins who are also members of a team with lower
permissions.

In the revised model, return both the aggregate permission and
additional lower permissions that are granted directly so that they may
still be used for reviewer assignment. However, for the approval to
actually count, the rule must specify the minimum permission and all
higher permissions that selected users may have.

For example, if a team has 'maintain' permission and a member of the
team is an org admin, that member's approval will only count if the rule
requires both 'admin' and 'maintain' permission.