Use minimum permission for approval checks
Change the 'permissions' field to accept any approval from a user with
at least the minimum permission specified in the list. This means that
for approval purposes, there is never a reason to specify more than one
permissions.

For reviewer assignment, it may be useful to specify more than one
permission to expand the set of users who are eligible for selection.
For example, this would allow selecting reviews from both admin and
maintainer teams.