Return 404 error when policy-bot is not installed
Previously, the server returned a 500 error with no details when trying
to view details for a PR where policy-bot was not installed. This was
most likely to happen if people followed existing links in a repository
that was moved between organizations.

Ideally, this would use a nice HTML 404 page, but for now, just include
a more descriptive plain text message.