Skip reviewer assignment on comment/review events
With the current set of predicates, these events can never change the
set of requested reviewers or trigger a re-request so there's no reason
to do the work of computing and assigning reviewers.

This is an alternate solution to the re-request loops that I fixed by
including head commit reviews in the set of reviewers. I think the other
approach is more correct, but this doesn't hurt.

The main downside is that if we add predicates that use comments or
reviews in the future, it will be easy to forget to change this trigger
condition and it might be a while before someone reports unexpected
behavior with the new predicate and reviewer assignment.