Add support for shared organization policies
Refactor configuration loading to use the go-githubapp/appconfig
package. This mostly adds support for shared organization policies (in
the .github repository by default), but should also make error messages
more accurate.

It also changes the policy link in the details view to point to the
resolved policy, instead of the local policy file.