Enable GraphQL query logging (#391)
Because v3 API request logs are not optional (other than by setting the
log level to something higher than debug), enable GraphQL query logs by
default. We can look at making this optional in the future if there is a
need.

Also upgrades google/go-github to v43, as a requirement of upgrading
palantir/go-githubapp to v0.12.0.