Add EvalContext to reduce handler duplication
Previously, we defined a bunch of common functionality for evaluating
PRs on the Base handler. This worked, but had a few flaws:

* When a handler needed to drop to a lower level to insert logic between
  steps, there was a lot of client construction boilerplate. This is not
  totally eliminated, but is reduced.

* As new side-effects are added, it was easy to miss calling them from
  places that had dropped down to the lower-level functions. For
  example, when adding review dismissal, we missed called it from the
  comment handler.

* Functions took a lot of similar arguments that could be defined on a
  single struct and shared.

* There was no easy way to evalute a PR without posting status check

* The Base handler was getting hard to read with all the evaluation
  functionality mixed in.

The EvalContext type contains all the configuration and clients needed
to evaluate a single PR for a request. Most handlers call Base.Evaluate
like before, but when dropping to a lower level, handlers now create and
use an EvalContext. Once created, they call 3 methods in sequence adding
whatever extra logic is needed in between.