Retry commit loading when pushed dates are missing
Back in 1d06dfa771, I removed retries for pushed date loading,
preferring to always use a fallback API. This seemed to work better in
most cases, but we're still seeing persistent but temporary failures
loading pushed dates for certain merge commits.

Add retries for the whole process again, covering both the initial pull
request API and the fallback commit APIs, in the hope that this gives
enough time for GitHub to update the internal data.