Implement invalidate_on_push without pushedDate field (#602)
Approximate the pushed time for commits by looking for the creation time
of the first status check on the commit. Because GitHub will only notify
integrations via webhook after a push, this time is guaranteed to be
after the actual push time. If no statuses exist, use the current time
as the push time.

In normal use, this should work well, because the webhook delivery and
processing latency is only a few seconds, but it could cause problems if
things slow down. In this case, Policy Bot fails close and will reject
approvals that technically could be accepted. This may be annoying, but
will only happen if users try to approve changes immediately after
pushes, which is hopefully rare.