Fix cache behavior for commits without statuses
This required a bigger change than expected because we want to cache the
evaluation timestamp when we use it. This meant moving the batch
checking in to the PushedAt function instead of handling it in the
approval logic. The new version is closer to how things used to work
when we loaded the pushedDate field.