Add global cache for pushed times (#612)
After testing, I'm worried that listing statuses on busy repositories
will cause too many API requests, even with HTTP caching. Since these
values are safe to cache forever, add a simple in-memory LRU cache in
each server instance to minimize API calls.

This required addional changes to how the PushedAt function works
because we want to cache the evaluation timestamp if we use it. This
meant moving the batch checking in to the PushedAt function instead of
handling it in the approval logic. The new version is closer to how
things used to work when we loaded the pushedDate field from the API.

Also add a test for the PushedAt context function that verifies the
caching behavior.