Add option to expand reviewers in details UI (#656)
If the `options.expand_required_reviewers` server option is set, when a
user expands a rule in the details UI, policy-bot will expand the teams,
organizations, and permission levels that can approve that rule in to a
list of user names. This can make it easier to figure out who needs to
approve your PR.

This feature is disabled by default because it can expose otherwise
private information about membership. Only enable it in trusted
environments.

This uses the HTMX library to handle loading user lists in the frontend
on demand, instead of precomputing all of the information everytime
someone views the details of a PR.