Simplify "hide skipped rules" implementation (#658)
It's always bothered me that the "hide skipped rules" toggle had to
remove elements from the DOM to hide them so that the :last-child
selector would work. This was fine when the toggle was off by default,
but became a problem once we switched it to default on: skipped rules
appear and then suddenly disappear once the DOM loads and the script
runs.

The new version does all of the hiding with CSS rules, so the JS toggle
only flips an attribute value instead adding or removing nodes. While
this needed a bit of server support in the template, it should make the
loading experience better and the toggle faster with large policies. And
because the tree is otherwise static, there's no disadvantage to
attaching extra information for styling in the template.