Load pull request draft status from GraphQL API (#742)
Back in 2019, we disabled this field because it was part of a GraphQL
preview. As a result, IsDraft was always false for PRs that we loaded
via the GraphQL query. For the most part, this doesn't matter, but can
cause problems with reviewer requests triggered via webhooks (like
status events) that do not provide complete pull request details.

This field has been out of preview for a long time now and is available
in all supported versions of GitHub, so it should be safe to load.