feat(predicate): Add `has_workflow_result` predicate
This is a new predicate which matches on the results of an entire GitHub
Actions workflow. This is similar to `has_status`, except that matches
on statuses only, which for GitHub Actions are roughly equivalent to
_jobs_ in a Workflow. It's often convenient to match on every job.

The workflows are given as paths in the repository, as these can't be
dynamic and are easy to predict when writing policies.