feat(policy): Support marshalling of config
Sometimes it's convenient to generate a Policy Bot config. For example,
when handling required but optional workflows, you need the config file to
contain a rule for each workflow which replicates its path filters.

That job is made much easier if the Policy Bot types can be reused. But
without `omitempty` the generated YAML contains all fields and it's hard
to read/review.

Add `omitempty` where it's needed, and `IsZero` in a few places too -
which is how the behaviour of `omitempty` can be customised.