Fix false context matches in status processing
Policy Bot looks for its own status context to identify when people or
apps might be impersonating it, but it used a simple prefix match. This
meant it could falsely trigger on unrelated contexts that shared the
same prefix. Change to using an exact match or a prefix match followed
by a ':', which will match the contexts with a branch name.