Add `has_status` predicate (#789)
The `has_successful_status` predicate currently requires predicates to
be present and successful in order to pass. But workflows can be run
conditionally - for example only if certain paths change - and it is
currently not very convenient to write a policy which considers such
skipped workflows as passing. The only feasible workaround is to
duplicate the path filters in the policy, and this quickly gets unwieldy
and prone to getting out-of-sync in large repositories.

Here we add direct support for specifying such rules. This is done by
introducing a new `has_status` predicate that accepts different
conclusions:

    has_status:
      conclusions: ["success", "skipped"]
      statuses:
        - "status 1"
        - "status 2"

Also mark the `has_successful_status` predicate as deprecated, since
`has_status` can do the same thing (and is equivalent when there are no
conclusions specified).