Don't look at workflow run results in other repositories
This is a copy of 536d20206c069699863acb75054283725b023c45. The same
issue exists with the `workflow_run` event as with `check_run`. People
can fork a repository and then propose pull requests from the original
repository into their fork. This may result in webhooks being
dispatched, but we want to ignore them.