Add accessor methods for Methods properties
Most of these can already be nil, removing the need to add pointers to
distinguish between set, set to non-zero, and set to zero values.
However, the accessors are still useful to allow getting default values
in a future PR. It also helps in the one case of a boolean pointer where
we had to check for nil each time.